minor reformatting draft for non macro polling

+++ lib/i2c_common/i2c_common.h
@@ -36,7 +36,7 @@ typedef enum i2c_mode_t {
     TWI_slave = 1
 } i2c_mode_t;
 
-typedef uint16_t i2c_return_t;
+typedef uint8_t i2c_return_t;
 
 i2c_return_t TWI_init(uint32_t frequency, uint8_t address);
 i2c_return_t TWI_check_status(i2c_status_t operation);
